Reuse scalar_as_i64 for Trilu and simplify Einsum and Attention helpers
scalar_as_i64 no longer casts a value that is already i64, which removes an
unnecessary_cast from every caller's generated code. burn_equation drops
its early return, which only repeated the rewrite's own condition, and
Attention's static_dim uses ArgType::static_shape.

+++ crates/burn-onnx/src/import/burn/node/div.rs
@@ -348,7 +348,7 @@ mod tests {
         pub fn forward(&self, lhs: [i64; 4], rhs: i64) -> [i64; 4] {
             let output = {
                 let mut result = lhs;
-                let scalar = rhs as i64;
+                let scalar = rhs;
                 for result_item in result.iter_mut() {
                     *result_item = if scalar != 0 {
                         *result_item / scalar
@@ -374,7 +374,7 @@ mod tests {
         pub fn forward(&self, lhs: [i64; 4], rhs: Tensor<1, Int>) -> [i64; 4] {
             let output = {
                 let mut result = lhs;
-                let scalar = (rhs).into_scalar::<i64>() as i64;
+                let scalar = (rhs).into_scalar::<i64>();
                 for result_item in result.iter_mut() {
                     *result_item = if scalar != 0 {
                         *result_item / scalar
@@ -400,7 +400,7 @@ mod tests {
         pub fn forward(&self, lhs: i64, rhs: [i64; 4]) -> [i64; 4] {
             let output = {
                 let mut result = rhs;
-                let scalar = lhs as i64;
+                let scalar = lhs;
                 for result_item in result.iter_mut() {
                     *result_item = if *result_item != 0 {
                         scalar / *result_item
@@ -426,7 +426,7 @@ mod tests {
         pub fn forward(&self, lhs: Tensor<1, Int>, rhs: [i64; 4]) -> [i64; 4] {
             let output = {
                 let mut result = rhs;
-                let scalar = (lhs).into_scalar::<i64>() as i64;
+                let scalar = (lhs).into_scalar::<i64>();
                 for result_item in result.iter_mut() {
                     *result_item = if *result_item != 0 {
                         scalar / *result_item
