+++ crates/burn-onnx/src/import/burn/node/modulo.rs
@@ -667,7 +667,7 @@ mod tests {
         pub fn forward(&self, lhs: [i64; 3], rhs: i64) -> [i64; 3] {
             let output = {
                 let mut result = lhs;
-                let scalar = rhs as i64;
+                let scalar = rhs;
                 for result_item in result.iter_mut() {
                     *result_item %= scalar;
                 }
@@ -693,7 +693,7 @@ mod tests {
         pub fn forward(&self, lhs: i64, rhs: [i64; 3]) -> [i64; 3] {
             let output = {
                 let mut result = rhs;
-                let scalar = lhs as i64;
+                let scalar = lhs;
                 for result_item in result.iter_mut() {
                     *result_item = scalar % *result_item;
                 }
@@ -827,7 +827,7 @@ mod tests {
         pub fn forward(&self, lhs: [i64; 3], rhs: i64) -> [i64; 3] {
             let output = {
                 let mut result = lhs;
-                let scalar = rhs as i64;
+                let scalar = rhs;
                 for result_item in result.iter_mut() {
                     *result_item %= scalar;
                 }
@@ -851,7 +851,7 @@ mod tests {
         pub fn forward(&self, lhs: i64, rhs: [i64; 3]) -> [i64; 3] {
             let output = {
                 let mut result = rhs;
-                let scalar = lhs as i64;
+                let scalar = lhs;
                 for result_item in result.iter_mut() {
                     *result_item = scalar % *result_item;
                 }
@@ -879,7 +879,7 @@ mod tests {
         pub fn forward(&self, lhs: [i64; 3], rhs: Tensor<1, Int>) -> [i64; 3] {
             let output = {
                 let mut result = lhs;
-                let scalar = (rhs).into_scalar::<i64>() as i64;
+                let scalar = (rhs).into_scalar::<i64>();
                 for result_item in result.iter_mut() {
                     *result_item %= scalar;
                 }
@@ -903,7 +903,7 @@ mod tests {
         pub fn forward(&self, lhs: Tensor<1, Int>, rhs: [i64; 3]) -> [i64; 3] {
             let output = {
                 let mut result = rhs;
-                let scalar = (lhs).into_scalar::<i64>() as i64;
+                let scalar = (lhs).into_scalar::<i64>();
                 for result_item in result.iter_mut() {
                     *result_item = scalar % *result_item;
                 }
